metralabs_python/requirements.txt
addict==2.4.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
cycler==0.12.1
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
fonttools==4.51.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
ipython==8.24.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
munkres==1.1.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
open3d==0.18.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
pluggy==1.5.0
pooch==1.8.1
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyquaternion==0.9.9
pytest==8.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
pyvista==0.43.7
pyvistaqt==0.11.0
PyYAML==6.0.1
QtPy==2.4.1
referencing==0.35.1
requests==2.31.0
retrying==1.3.4
rpds-py==0.18.1
scikit-learn==1.4.2
scipy==1.13.0
scooby==0.10.0
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
threadpoolctl==3.5.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vtk==9.3.0
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.10
zipp==3.18.1
